refactor: (#175, #177) 학교 조회 API 분리
